ml-api/requirement.txt
tensorflow==2.12.0
flask==2.3.2
pillow==10.0.0
numpy==1.24.3
gunicorn==20.1.0
flask-cors==3.0.10
imageio==2.31.1
